grant multi-category filter
