+++ src/components/Category/CategoryOption.tsx
@@ -1,32 +1,26 @@
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback } from 'react'
 
 import classNames from 'classnames'
 import isNumber from 'lodash/isNumber'
 import toSnakeCase from 'lodash/snakeCase'
 
-import { NewGrantCategory, SubtypeAlternativeOptions, SubtypeOptions, toGrantSubtype } from '../../entities/Grant/types'
-import { ProposalType } from '../../entities/Proposal/types'
+import { NewGrantCategory } from '../../entities/Grant/types'
 import { CategoryIconVariant } from '../../helpers/styles'
 import useFormatMessage from '../../hooks/useFormatMessage'
-import useURLSearchParams from '../../hooks/useURLSearchParams'
 import Link from '../Common/Typography/Link'
 import Text from '../Common/Typography/Text'
-import Arrow from '../Icon/Arrow'
 import { getNewGrantsCategoryIcon } from '../Icon/NewGrantsCategoryIcons'
 import All from '../Icon/ProposalCategories/All'
 import Grant from '../Icon/ProposalCategories/Grant'
 import Tender from '../Icon/ProposalCategories/Tender'
-import SubItem from '../Icon/SubItem'
-import { ProjectTypeFilter } from '../Search/CategoryFilter'
 
 import { categoryIcons } from './CategoryBanner'
 import './CategoryOption.css'
 
-type Props = Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, 'children'> & {
+export type CategoryOptionProps = Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, 'children'> & {
   active?: boolean
   type: string
   count?: number
-  subtypes?: SubtypeOptions[]
   onClick?: () => void
 }
 
@@ -61,22 +55,8 @@ export const getCategoryIcon = (type: string, variant?: CategoryIconVariant, siz
   )
 }
 
-const getHref = (href: string | undefined, subtype: SubtypeOptions) => {
-  const url = new URL(href || '/', 'http://localhost') // Create a URL object using a dummy URL
-  if (subtype === SubtypeAlternativeOptions.All) {
-    url.searchParams.delete('subtype')
-  } else {
-    url.searchParams.set('subtype', subtype)
-  }
-  const newHref = url.pathname + '?' + url.searchParams.toString()
-  return newHref
-}
-
-export default function CategoryOption({ active, type, className, count, subtypes, onClick, ...props }: Props) {
+export default function CategoryOption({ active, type, className, count, onClick, ...props }: CategoryOptionProps) {
   const t = useFormatMessage()
-  const params = useURLSearchParams()
-  const currentType = useMemo(() => params.get('type'), [params])
-  const currentSubtype = useMemo(() => toGrantSubtype(params.get('subtype')), [params])
 
   const handleClick = useCallback(
     (e: React.MouseEvent<HTMLAnchorElement>) => {
@@ -85,25 +65,6 @@ export default function CategoryOption({ active, type, className, count, subtype
       }
     },
     [onClick]
-  )
-
-  const isGrant = currentType === ProposalType.Grant || currentType === ProjectTypeFilter.Grants
-  const hasSubtypes = !!subtypes && subtypes.length > 0
-  const [isSubtypesOpen, setIsSubtypesOpen] = useState(isGrant)
-
-  useEffect(() => {
-    setIsSubtypesOpen(isGrant)
-  }, [isGrant, currentType])
-
-  const isSubtypeActive = useCallback(
-    (subtype: SubtypeOptions) => {
-      if (subtype === SubtypeAlternativeOptions.All && !currentSubtype) {
-        return true
-      }
-
-      return toSnakeCase(subtype) === toSnakeCase(currentSubtype)
-    },
-    [currentSubtype]
   )
 
   return (
@@ -124,17 +85,6 @@ export default function CategoryOption({ active, type, className, count, subtype
             {getCategoryIcon(type, CategoryIconVariant.Circled)}
             <Text className="CategoryOption__Title">{t(`category.${type}_title`)}</Text>
           </span>
-          {/* {hasSubtypes && (
-            <span
-              className={classNames('CategoryOption__Arrow', isSubtypesOpen && 'CategoryOption__Arrow--active')}
-              onClick={(e) => {
-                e.preventDefault()
-                setIsSubtypesOpen((prev) => !prev)
-              }}
-            >
-              <Arrow filled={false} color="var(--black-700)" />
-            </span>
-          )} */}
         </span>
         {isNumber(count) && (
           <span className={classNames('CategoryOption__Counter', active && 'CategoryOption__Counter--active')}>
@@ -142,30 +92,6 @@ export default function CategoryOption({ active, type, className, count, subtype
           </span>
         )}
       </Link>
-      {/* {hasSubtypes && (
-        <div
-          className={classNames(
-            'CategoryOption__Subcategories',
-            isSubtypesOpen && 'CategoryOption__Subcategories--active'
-          )}
-        >
-          {subtypes.map((subtype, index) => {
-            return (
-              <Link
-                className={classNames(
-                  'CategoryOption__SubcategoryItem',
-                  isSubtypeActive(subtype) && 'CategoryOption__SubcategoryItem--active'
-                )}
-                key={subtype + `-${index}`}
-                // href={getHref(props.href, subtype)}
-              >
-                <SubItem />
-                {t(`category.${toSnakeCase(subtype)}_title`)}
-              </Link>
-            )
-          })}
-        </div>
-      )} */}
     </>
   )
 }
